rework the way we enforce a 'minimum' log level.
doing it the previous way was causing issues with the initial log level
being set correctly, as the logger delegate's starting level is often
'null' (which means it defers to its parent).  we don't know how that
parent level compares with the enforced minimum and finding it involves
walking the parent tree until we find an actual level.  instead, just
enforce it in isLoggable.

also, rename maxLogLevel to minLogLevel, as that's more consistent with
how the log level values are defined.