Logger v2 (#11)
* wip: new logger classes/logger context

* logcontext tweaks/tests

* fleshing out new logger, start of tests

* add var for start/end context tokens, expose test helpers

* don't include a context chunk if the context is empty

* tweaks to logger and some basic tests

* fix a bug with setting log level

* move new logging files to logging2 package

* rename logger classes

* change the way we create a subcontext

* add another overload for creating a child logger without additional context

* change the way we pass the log context data to the logger

* add a log formatter

* add throwable versions for warn and error

* rework the way we enforce a 'minimum' log level.

doing it the previous way was causing issues with the initial log level
being set correctly, as the logger delegate's starting level is often
'null' (which means it defers to its parent).  we don't know how that
parent level compares with the enforced minimum and finding it involves
walking the parent tree until we find an actual level.  instead, just
enforce it in isLoggable.

also, rename maxLogLevel to minLogLevel, as that's more consistent with
how the log level values are defined.

* only add context if it's non-empty

* add copyrights

* tweak log format

* another log format tweak

* better handle old logger

* fix docs in logger interface

* add missing copyright