move read frequency/optinal-ity out to strategies
i don't think we can get away without the idea of an optional
property.  introducing subclasses of {ReadOnly|ReadEveryTime}Property
would lead to a large amount of classes, so i've refactored 'read
frequency' out to a 'strategy' which is applied to a config property.
'prop not found' is another strategy which defines what behavior should
occur when a property isn't found (so in the case of an 'optional'
config property we can just return null)