use SPI instead of a fixed config interface
this means that code can request retrieving any type it wants and, as
long as a service provider provides that type, it will retrieve it.  in
practice, this means that we can take advantage of more complex features
like retrieving a ConfigObject from typesafe config when we need to do
more complex work when retrieving a property,  Otherwise, we would have
had to implement a common-mirrored type to something like ConfigObject
(as well as the conversion logic).  Now, rather than boxing
functionality from any possible config source into some lowest-common
denominator, we can use them directly.