Configk (#27)
* config meta lib in kotlin

* cleanup, prop attributes, support for optional props

* add support for duration

* refactor typed prop delegate code

* refactor how optionals are done

* rip out optional

with kotlin, optional--or, at least, using nullable to represent
optional--just doesn't seem to work, due to the fact that it's
impossible to have a top-level config prop type since we'd need to
either have the value be a T (required) or T? (optional).

if we do need to bring back optional, we'll have to do it in a way other
than exposing T or T? (like maybe just adding a separate 'valueOrNull'
member in addition to 'value', which can swallow the not-found exception
and return null instead

* cleanup...this feels like it's on the right track

* use custom ConfigResult instead of Kotlin's result

Kotlin's result doesn't appear to give a way to get at the actual,
unwrapped type (?!?).  either way, it's not technically stable yet and
it seems they suggest going the sealed class route for
this use case.

* more tests

* create a sources jar

* better exceptions

* use SPI instead of a fixed config interface

this means that code can request retrieving any type it wants and, as
long as a service provider provides that type, it will retrieve it.  in
practice, this means that we can take advantage of more complex features
like retrieving a ConfigObject from typesafe config when we need to do
more complex work when retrieving a property,  Otherwise, we would have
had to implement a common-mirrored type to something like ConfigObject
(as well as the conversion logic).  Now, rather than boxing
functionality from any possible config source into some lowest-common
denominator, we can use them directly.

* cleanup/docs

* fix test dir structure

* ReadStrategy -> ReadFrequencyStrategy

* dsl take 1

* add a prop value retriever to do the retrieval 'heavy lifting'

this used to be done in the property delegate, but we sometimes don't
want to use a delegate, so it makes more sense for the delegate to be a
thin wrappe over a retriever

* more dsl experimentation

* remove prop delegate

* a multitude of attempts at getting a better property definition syntax

think i've finally gotten one that feels the best in BuilderTest.kt

* get rid of the need for SPI to retrieve typed values

define a ConfigSource interface instead with this method since we now
pass the config source around in attributes we can retrieve it from there directly

* combine property attrs builders

* improve retrievedAs syntax

* add examples

* add multi-property helpers

* tweak exampleconfigsource for examples

* add example simple prop helper

* clean up config source api/improve errors

* clean out old code

* move configk2 package to configk

* add equals/hash/toString for configresult types

* add read frequency strategy tests

* docs

* property tests, fix bug in converting prop where we invoked the converter every time, even if it was read-once

* more tests/examples

* add support for deprercation attribute

* simplify pasthrough code in prop builder by using attr builder class
delegate

* docs

* add name to config source, docs, tweak exception toString

* minor tweaks

* more attrs docs

* remove old file

* fix bug in getOrElse, add more docs

* add a readme

* delete old config

* configk -> config

* delete old config tests