change config properties to center around attributes rather than anonymous-class instances.
creating anonymous class instances allows for a very slick syntax at the
definition site, but it doesn't result in very testable properties as we
can't easily re-instantiate them with different underying configuration
setups for testing.  now, we move the helpers/builders to be centered
around attributes to make building them easy when defining a property
subclass.