fix: Slow initial convergence in RateTracker
We used to use the full window as the denumerator when calculating the average rate.
This results in slow convergence at the start, e.g. a tracker with a 5 s
window would return only 20% of the observed rate when queried 1 s after
a value was added.

We now keep track of how much of the window is used, so we have a more accurate rate earlier.